% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{CIViC_API}
\alias{CIViC_API}
\alias{evidence}
\alias{evidence_comments}
\alias{evidence_detail}
\alias{evidence_suggested_changes}
\alias{evidence_suggested_changes_comments}
\alias{gene_detail}
\alias{genes}
\alias{genes_comments}
\alias{variant_detail}
\alias{variant_groups}
\alias{variant_groups_comments}
\alias{variants}
\alias{variants_comments}
\title{Accessing CIViC resources}
\usage{
genes(page = 1, count = 25)

gene_detail(id, identifier_type = c("civic_id", "entrez_id", "entrez_symbol"))

genes_comments(id)

variants(page = 1, count = 25)

variant_detail(id)

variants_comments(id)

evidence(page = 1, count = 25)

evidence_detail(id)

evidence_comments(id)

evidence_suggested_changes(id)

evidence_suggested_changes_comments(id, cid)

variant_groups(page = 1, count = 25)

variant_groups_comments(id)
}
\description{
Accessing CIViC resources
}

